doc: Better demo layout
