+++ packages/docs/content/docs/parsers.mdx
@@ -14,7 +14,7 @@ import {
   DateISOParserDemo,
   DateTimestampParserDemo,
   CustomParserDemo
-} from './parsers-demos'
+} from './demos'
 
 Search params are strings by default, but chances are your state is more complex than that.
