+++ packages/docs/src/app/playground/(demos)/_components/query-spy.skeleton.tsx
@@ -10,7 +10,7 @@ export function QuerySpySkeleton({
       aria-label="Querystring spy"
       aria-description="For browsers where the query is hard to see (eg: on mobile)"
       className={twMerge(
-        'mt-4 block w-full overflow-x-auto rounded-lg border bg-background px-3 py-2 text-xs dark:bg-zinc-900 dark:shadow-inner sm:text-sm',
+        'mt-4 block w-full overflow-x-auto rounded-lg border bg-background px-3 py-2 text-xs dark:bg-zinc-900/50 dark:shadow-inner sm:text-sm',
         className
       )}
       {...props}
